python/requirements.txt
python-dotenv==1.0.1
redis==5.0.3
autopep8==2.1.0
flake8==7.0.0
